requirements.txt
streamlit>=1.30.0
gitpython>=3.1.40
pandas>=2.1.0
numpy>=1.26.0
matplotlib>=3.8.0
seaborn>=0.13.0
transformers>=4.36.0
torch>=2.1.0
requests>=2.31.0
